+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This file is used to list changes made in each version of cookbook-openstack-orc
 * Add attributes for stack role and domain users
 * Bump Chef gem to 11.16
 * Create role and domain setup for heat template defined users
+* Add cert_file, key_file, ca_file and insecure for clients so that they are configurable.
 
 ## 9.2.0
 * python_packages database client attributes have been migrated to